+++ src/formElementsService.ts
@@ -213,7 +213,7 @@ function flattenFormElements(
 }
 
 /** The regex used for matching `{ELEMENT:<elementName>}` tags in the OneBlink platform. */
-const WYSIWYGRegex = /({ELEMENT:)([^}]+)(})/g
+const ElementWYSIWYGRegex = /({ELEMENT:)([^}]+)(})/g
 
 /**
  * Takes a string and calls a provided handler function for each found instance
@@ -244,11 +244,47 @@ const matchElementsTagRegex = (
   }) => void,
 ) => {
   let matches
-  while ((matches = WYSIWYGRegex.exec(data)) !== null) {
+  while ((matches = ElementWYSIWYGRegex.exec(data)) !== null) {
     if (matches?.length < 3) continue
 
     const elementName = matches[2]
     matchHandler({ elementName, elementMatch: matches[0] })
+  }
+}
+
+/** The regex used for matching the `{USER:email}` tag in the OneBlink platform. */
+const UserWYSIWYGRegex = /{USER:email}/g
+
+/**
+ * Takes a string and calls a provided handler function for each found instance
+ * of `{ELEMENT:<elementName>}` in the string. Used to replace values in
+ * OneBlink calculation and info (HTML) elements.
+ *
+ * #### Example
+ *
+ * ```js
+ * formElementsService.matchElementsTagRegex(
+ *   myString,
+ *   ({ elementName, elementMatch }) => {
+ *     const v = submission[elementName]
+ *     myString = myString.replace(elementMatch, v)
+ *   },
+ * )
+ * ```
+ *
+ * @param data
+ * @param handler
+ * @returns
+ */
+const matchUserTagRegex = (
+  data: string,
+  matchHandler: (userMatch: string) => void,
+) => {
+  let matches
+  while ((matches = UserWYSIWYGRegex.exec(data)) !== null) {
+    if (matches?.length < 1) continue
+
+    matchHandler(matches[0])
   }
 }
 
@@ -258,6 +294,8 @@ export {
   findFormElement,
   parseFormElementOptionsSet,
   flattenFormElements,
-  WYSIWYGRegex,
+  ElementWYSIWYGRegex,
+  UserWYSIWYGRegex,
   matchElementsTagRegex,
+  matchUserTagRegex,
 }