+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - `formElementsService.determineIsInfoPage()`
+- `gov-pay` payment event to payment event check
 
 ## [3.0.0] - 2023-06-05